Make some UUID utilities public (fixes #90)
